Muevo operaciones de indexación a un módulo aparte
    refs:#107
